refactor(umbp): share host registration and gather helpers
Port of upstream 93f2998a. Moves HostTierRegistration and LaunchDeviceGather
from DRAM-tier-private headers into src/umbp/include/umbp/common/, leaving
forwarding shims at the old paths so the tier and its tests keep compiling
unchanged. The distributed ranged-I/O work needs both helpers outside the
local tier, which is what this move is for.

Deviation from upstream: upstream's move rewrote the two headers' comments
down to a few lines. This keeps the original rationale — the measured
per-fragment costs that justify the kernel over hipMemcpy/hipMemcpy2DAsync,
the ~128 KiB crossover, and why registration is single-shot and
all-or-nothing — because those numbers are why the code is shaped this way.
