+++ src/umbp/local/tiers/host_registration.cpp
@@ -19,7 +19,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-#include "host_registration.h"
+#include "umbp/common/host_registration.h"
 
 #include <hip/hip_runtime_api.h>
